【Feature】flutter_network、flutter_network_kti、app_network

+++ flutter_network_kit/lib/src/base/cache_network_client.dart
@@ -38,12 +38,18 @@ class CacheNetworkClient extends NetworkClient {
   /// delete local cache by primaryKey and optional subKey
   Future<bool> delete(String primaryKey,
       {String? requestMethod, String? subKey}) {
+    if (_dioCacheManager == null) {
+      return Future.value(true);
+    }
     return _dioCacheManager!
         .delete(primaryKey, requestMethod: requestMethod, subKey: subKey);
   }
 
   /// no matter what subKey is, delete local cache if primary matched.
   Future<bool> deleteByPrimaryKeyWithUri(Uri uri, {String? requestMethod}) {
+    if (_dioCacheManager == null) {
+      return Future.value(true);
+    }
     return _dioCacheManager!
         .deleteByPrimaryKeyAndSubKeyWithUri(uri, requestMethod: requestMethod);
   }
@@ -51,6 +57,9 @@ class CacheNetworkClient extends NetworkClient {
   /// delete local cache when both primaryKey and subKey matched.
   Future<bool> deleteByPrimaryKeyAndSubKeyWithUri(Uri uri,
       {String? requestMethod, String? subKey, dynamic data}) {
+    if (_dioCacheManager == null) {
+      return Future.value(true);
+    }
     return _dioCacheManager!.deleteByPrimaryKeyAndSubKeyWithUri(uri,
         requestMethod: requestMethod, subKey: subKey, data: data);
   }
@@ -65,7 +74,9 @@ class CacheNetworkClient extends NetworkClient {
     required Map<String, dynamic> headerCommonFixParams, // header 中公共但不变的参数
     Map<String, dynamic> Function()?
         headerCommonChangeParamsGetBlock, // header 中公共但会变的参数
-    required String headerAuthorization,
+    String? headerAuthorization,
+    List<String>?
+        headerAuthorizationWhiteList, // 请求时候，检查到当前headerAuthorization为空,可继续请求的白名单
     required Map<String, dynamic> bodyCommonFixParams, // body 中公共但不变的参数
     Map<String, dynamic> Function()?
         bodyCommonChangeParamsGetBlock, // body 中公共但会变的参数
@@ -87,6 +98,7 @@ class CacheNetworkClient extends NetworkClient {
       headerCommonFixParams: headerCommonFixParams,
       headerCommonChangeParamsGetBlock: headerCommonChangeParamsGetBlock,
       headerAuthorization: headerAuthorization,
+      headerAuthorizationWhiteList: headerAuthorizationWhiteList,
       bodyCommonFixParams: bodyCommonFixParams,
       bodyCommonChangeParamsGetBlock: bodyCommonChangeParamsGetBlock,
       dealRequestOptionsAction: (RequestOptions options) {
@@ -115,117 +127,92 @@ class CacheNetworkClient extends NetworkClient {
   }
 
   /// 通用的POST请求(如果设置缓存，可实现如果从缓存中取到数据，仍然能继续执行正常的请求)
-  /*
-  Future<ResponseModel> _post(
-    String api, {
-    Map<String, dynamic> customParams,
-    NetworkCacheLevel cacheLevel = NetworkCacheLevel.none,
-  }) async {
-    Completer<ResponseModel> completer = Completer();
-    postWithCallback(
-      api,
-      customParams: customParams,
-      cacheLevel: cacheLevel,
-      completeCallBack: (responseModel) {
-        if (responseModel.isCache != true) {
-          completer.complete(responseModel);
-        }
-      },
-    );
 
-    return completer.future;
-  }
-  */
-
-  /// 具有更丰富功能的 post 接口(因为即使有缓存也只会返回一次，所以无法使用)
-  Future<ResponseModel> cache_post(
+  /// 可进行缓存的请求(需要多次返回结果)
+  void cache_requestWithCallback(
     String api, {
+    RequestMethod requestMethod = RequestMethod.post,
     Map<String, dynamic>? customParams,
-    int retryCount = 0,
-    NetworkCacheLevel cacheLevel = NetworkCacheLevel.none,
-  }) async {
-    Completer _requestCompleter = Completer<String>();
-
-    late ResponseModel responseModel;
-    cache_postWithCallback(
-      api,
-      customParams: customParams,
-      retryCount: retryCount,
-      cacheLevel: cacheLevel,
-      completeCallBack: (ResponseModel bResponseModel) {
-        if (_requestCompleter.isCompleted != true) {
-          _requestCompleter.complete('所有请求(含重试、缓存)结束');
-        }
-
-        responseModel = bResponseModel;
-      },
-    );
-
-    await _requestCompleter.future;
-    return responseModel;
-  }
-
-  void cache_postWithCallback(
-    String api, {
-    Map<String, dynamic>? customParams,
+    bool? ifNoAuthorizationForceGiveUpRequest, // 没有 Authorization 的时候是否强制放弃请求
     int retryCount = 0,
     NetworkCacheLevel cacheLevel = NetworkCacheLevel.none,
     required void Function(ResponseModel responseModel) completeCallBack,
-  }) {
-    postWithRetryCallback(
+    ResponseModel? cacheResponseModel,
+  }) async {
+    /*
+    bool shouldGiveUp = await super.shouldGiveUpRequest(api);
+    // 此时可能网络初始化还没完成(PS:初始化操作在base_requestUrl中 await _initCompleter.future;)
+    if (shouldGiveUp) {
+      LogApiUtil.logCancelApi(api);
+      return;
+    }
+    */
+
+    if (ifNoAuthorizationForceGiveUpRequest == true) {
+      if (await existAuthorization() == false) {
+        return;
+      }
+    }
+
+    ResponseModel responseModel = await requestWithRetry(
       api,
+      requestMethod: requestMethod,
       customParams: customParams,
       retryCount: retryCount,
       options: CacheHelper.buildOptions(cacheLevel),
-      completeCallBack: (ResponseModel responseModel) async {
-        if (completeCallBack == null) {
-          debugPrint('温馨提示:本次请求，你没有实现回调方法$api');
-          return;
-        }
-
-        // 不是真正的网络请求返回的Response\Error结果(eg:比如是取缓存的结果时候)
-        bool noRealRequest = cacheLevel == NetworkCacheLevel.one;
-        if (responseModel.isCache != true) {
-          // 1、当请求结果是后台实际请求返回的时候:
-          if (responseModel.isSuccess) {
-            // ①.如果实际的请求成功，则直接返回
-            completeCallBack(responseModel);
-          } else {
-            // ②.如果实际的请求失败，由于 statusCode 在200到300之间的请求结果都会被保存起来，但我们此系统里他们可能是错误的。
-            // 如500，就不要保存起来了，防止请求的时候发现有缓存数据，使得页面使用到了保存的数据
-            if (cacheLevel != NetworkCacheLevel.none) {
-              Uri uri = getUri(api);
-              bool deleteSuccess = await deleteByPrimaryKeyAndSubKeyWithUri(uri,
-                  requestMethod: 'POST', data: customParams);
-              if (deleteSuccess == false) {
-                debugPrint("Warning:本不该纯，但存了，现在必须删却又没删除成功的接口");
-              }
-            }
-            completeCallBack(responseModel);
-          }
-        } else {
-          // 2、请求结果是之前缓存的数据返回的(即结果是真正的网络请求返回的Response\Error结果)的时候，
-          // ①先把缓存返回回去，
-          // ②再继续发起实际的请求，且新请求为后台实际请求并且其要会保存请求成功的数据
-          completeCallBack(responseModel);
-
-          late NetworkCacheLevel newCacheLevel;
-          if (cacheLevel == NetworkCacheLevel.one) {
-            newCacheLevel = NetworkCacheLevel.forceRefreshAndCacheOne;
-          } else {
-            debugPrint('Error：判断出错啦，此结果不是缓存数据，却走到了isCache==true');
-            newCacheLevel = NetworkCacheLevel
-                .none; //TODO:临时为了走下去，应该自始至终都不会走到这里，这里之后要 throw Exception
-          }
-
-          cache_postWithCallback(
-            api,
-            customParams: customParams,
-            cacheLevel: newCacheLevel,
-            completeCallBack: completeCallBack,
-          );
-        }
-      },
+      beforeResponseModel: cacheResponseModel,
     );
+
+    // 不是真正的网络请求返回的Response\Error结果(eg:比如是取缓存的结果时候)
+    bool noRealRequest = cacheLevel == NetworkCacheLevel.one;
+    if (responseModel.isCache != true) {
+      // 1、当请求结果是后台实际请求返回的时候:
+      if (responseModel.isSuccess) {
+        // ①.如果实际的请求成功，则直接返回
+        completeCallBack(responseModel);
+      } else {
+        // ②.如果实际的请求失败，由于 statusCode 在200到300之间的请求结果都会被保存起来，但我们此系统里他们可能是错误的。
+        // 如500，就不要保存起来了，防止请求的时候发现有缓存数据，使得页面使用到了保存的数据
+        if (cacheLevel != NetworkCacheLevel.none) {
+          Uri uri = getUri(api);
+          String requestMethodString = requestMethod.toString().split('.').last;
+          bool deleteSuccess = await deleteByPrimaryKeyAndSubKeyWithUri(uri,
+              requestMethod: requestMethodString, data: customParams);
+          if (deleteSuccess == false) {
+            _log("Warning:此份网络数据结果本不该存储，但存了，现在必须删却又没删除成功的接口");
+          }
+        }
+        completeCallBack(responseModel);
+      }
+    } else {
+      // 2、请求结果是之前缓存的数据返回的(即结果是真正的网络请求返回的Response\Error结果)的时候，
+      // ①先把缓存返回回去，
+      // ②再继续发起实际的请求，且新请求为后台实际请求并且其要会保存请求成功的数据
+      completeCallBack(responseModel);
+
+      late NetworkCacheLevel newCacheLevel;
+      if (cacheLevel == NetworkCacheLevel.one) {
+        newCacheLevel = NetworkCacheLevel.forceRefreshAndCacheOne;
+      } else {
+        _log(
+            'Error：${api}判断出错啦，此结果不是缓存数据，却走到了isCache==true'); // TODO:无网络的情况下会发生此现象
+        newCacheLevel = NetworkCacheLevel
+            .none; //TODO:临时为了走下去，应该自始至终都不会走到这里，这里之后要 throw Exception
+      }
+
+      cache_requestWithCallback(
+        api,
+        requestMethod: requestMethod,
+        customParams: customParams,
+        cacheLevel: newCacheLevel,
+        completeCallBack: completeCallBack,
+        cacheResponseModel: responseModel, // 缓存数据
+      );
+    }
+  }
+
+  _log(String message) {
+    String dateTimeString = DateTime.now().toString().substring(0, 19);
+    debugPrint('$dateTimeString:$message');
   }
 }
